refactor: replace watcher with action where needed
