+++ packages/clients/textLocator/src/plugins/GeometrySearch/store/index.ts
@@ -16,7 +16,7 @@ import { updateVectorLayer, vectorLayer } from '../utils/vectorDisplay'
 import { geoJson } from '../../../utils/coastalGazetteer/common'
 import { setupTooltip } from './actions/setupTooltip'
 import { setupDrawReaction } from './actions/setupDrawReaction'
-import { setupWatchers, updateFrequencies } from './actions/watchers'
+import { updateFrequencies } from './actions/updateFrequencies'
 
 let counter = 0
 const searchLoadingKey = 'geometrySearchLoadingKey'
@@ -41,14 +41,11 @@ export const makeStoreModule = () => {
         dispatch('setupDrawReaction')
         dispatch('setupTooltip')
         map.addLayer(vectorLayer)
-        // register watchers after store is ready (else immediate-like firing on not-really change)
-        setTimeout(() => dispatch('setupWatchers'), 0)
       },
       setupTooltip,
       setupDrawReaction,
-      setupWatchers,
       updateFrequencies,
-      searchGeometry({ rootGetters, commit }, feature: Feature) {
+      searchGeometry({ rootGetters, commit, dispatch }, feature: Feature) {
         const loadingKey = getSearchLoadingKey()
         commit('plugin/loadingIndicator/addLoadingKey', loadingKey, {
           root: true,
@@ -61,7 +58,10 @@ export const makeStoreModule = () => {
             rootGetters.configuration.geometrySearch.url,
             rootGetters.configuration.epsg
           )
-          .then((result) => commit('setFeatureCollection', result))
+          .then((result) => {
+            commit('setFeatureCollection', result)
+            dispatch('updateFrequencies')
+          })
           .finally(() =>
             commit('plugin/loadingIndicator/removeLoadingKey', loadingKey, {
               root: true,
