+++ packages/clients/textLocator/src/utils/coastalGazetteer/searchToponym.ts
@@ -58,7 +58,7 @@ export async function searchCoastalGazetteerByToponym(
 export const selectResult: SelectResultFunction<
   AddressSearchState,
   AddressSearchGetters
-> = ({ commit, rootGetters }, { feature }) => {
+> = ({ commit, dispatch, rootGetters }, { feature }) => {
   // default behaviour (AddressSearch selects and is not involved in further behaviour)
   commit('plugin/addressSearch/setChosenAddress', feature, { root: true })
   commit('plugin/addressSearch/setInputValue', feature.title, { root: true })
@@ -79,4 +79,5 @@ export const selectResult: SelectResultFunction<
     },
     { root: true }
   )
+  dispatch('plugin/geometrySearch/updateFrequencies', null, { root: true })
 }
